refactor: typesのuserフォルダを展開
