+++ src/components/auth/SignUpForm.tsx
@@ -7,7 +7,7 @@ import {
 import { RiLockPasswordLine as RiLockPasswordLineIcon } from 'react-icons/ri'
 import { Button } from 'components/shares/base/Button'
 
-import { SignUpRequestParams } from 'types/user/auth'
+import { SignUpRequestParams } from 'types/auth'
 
 type Props = {
   onSubmit: (params: SignUpRequestParams) => void
