+++ src/hooks/login/useAuth.ts
@@ -2,8 +2,8 @@ import { useRouter } from 'next/navigation'
 import { useCallback } from 'react'
 import { useGetApi } from 'hooks/useApi'
 import { useCookies } from 'stores/useCookies'
-import { LoginRequestParams, SignUpRequestParams } from 'types/user/auth'
-import { UserWithToken } from 'types/user/user'
+import { LoginRequestParams, SignUpRequestParams } from 'types/auth'
+import { UserWithToken } from 'types/user'
 import { HttpError, postApi } from 'utils/api'
 import { encrypted } from 'utils/encrypt'
